requirements.txt
Django==1.8.5
django-modelcluster==1.0
mysqlclient==1.3.6
pytz==2015.6
six==1.10.0
django-suit
django-material
djangorestframework